+++ src/main/java/ru/olegcherednik/zip4jvm/io/in/DataInput.java
@@ -76,17 +76,6 @@ public interface DataInput extends Marker, ReadBuffer, Closeable {
 
     long skip(long bytes) throws IOException;
 
-    default long seekForward(long destAbsOffs) throws IOException {
-        long absOffs = getAbsOffs();
-
-        if (destAbsOffs == absOffs)
-            return 0;
-        if (destAbsOffs < absOffs)
-            throw new IOException("can't move backward");
-
-        return skip(destAbsOffs - absOffs);
-    }
-
     default int readWordSignature() throws IOException {
         return readWord();
     }
