--- src/main/java/ru/olegcherednik/zip4jvm/io/in/file/consecutive/ConsecutiveDataInput.java
@@ -1,11 +1,0 @@
-package ru.olegcherednik.zip4jvm.io.in.file.consecutive;
-
-import ru.olegcherednik.zip4jvm.io.in.DataInput;
-
-/**
- * @author Oleg Cherednik
- * @since 20.12.2024
- */
-public interface ConsecutiveDataInput extends DataInput {
-
-}
