+++ src/main/java/ru/olegcherednik/zip4jvm/io/readers/ZipModelReader.java
@@ -18,7 +18,7 @@
  */
 package ru.olegcherednik.zip4jvm.io.readers;
 
-import ru.olegcherednik.zip4jvm.engine.UnzipEngine;
+import ru.olegcherednik.zip4jvm.engine.unzip.UnzipEngine;
 import ru.olegcherednik.zip4jvm.io.in.file.random.RandomAccessDataInput;
 import ru.olegcherednik.zip4jvm.io.readers.zip64.Zip64Reader;
 import ru.olegcherednik.zip4jvm.model.Charsets;
@@ -89,7 +89,7 @@ public final class ZipModelReader extends BaseZipModelReader {
 
     @Override
     protected RandomAccessDataInput createDataInput() {
-        return Quietly.doQuietly(() -> UnzipEngine.createDataInput(srcZip));
+        return Quietly.doQuietly(() -> UnzipEngine.createRandomAccessDataInput(srcZip));
     }
 
     @Override
