+++ src/main/java/ru/olegcherednik/zip4jvm/io/readers/block/BlockZipModelReader.java
@@ -18,7 +18,7 @@
  */
 package ru.olegcherednik.zip4jvm.io.readers.block;
 
-import ru.olegcherednik.zip4jvm.engine.UnzipEngine;
+import ru.olegcherednik.zip4jvm.engine.unzip.UnzipEngine;
 import ru.olegcherednik.zip4jvm.io.in.file.random.RandomAccessDataInput;
 import ru.olegcherednik.zip4jvm.io.readers.BaseZipModelReader;
 import ru.olegcherednik.zip4jvm.io.readers.CentralDirectoryReader;
@@ -88,7 +88,7 @@ public final class BlockZipModelReader extends BaseZipModelReader {
 
     @Override
     protected RandomAccessDataInput createDataInput() {
-        return Quietly.doQuietly(() -> UnzipEngine.createDataInput(srcZip));
+        return Quietly.doQuietly(() -> UnzipEngine.createRandomAccessDataInput(srcZip));
     }
 
     @Override
