+++ src/test/java/ru/olegcherednik/zip4jvm/ZipCompressionOptimizationTest.java
@@ -18,7 +18,7 @@
  */
 package ru.olegcherednik.zip4jvm;
 
-import ru.olegcherednik.zip4jvm.engine.InfoEngine;
+import ru.olegcherednik.zip4jvm.engine.info.InfoEngine;
 import ru.olegcherednik.zip4jvm.model.Compression;
 import ru.olegcherednik.zip4jvm.model.CompressionMethod;
 import ru.olegcherednik.zip4jvm.model.DataDescriptorEnum;
